Changed from static class to namespace
